% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bw2.R
\name{theme_bw2}
\alias{theme_bw2}
\title{theme_bw2}
\usage{
theme_bw2()
}
\value{

}
\description{
theme_bw2
}
\examples{
library(ggplot2)
ggplot2:: ggplot(iris, aes(Sepal.Length, Sepal.Width))+
 geom_point()+
 theme_bw2()

}
