% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_owid.R
\name{theme_owid}
\alias{theme_owid}
\title{theme_owid}
\usage{
theme_owid()
}
\value{

}
\description{
theme_owid
}
\examples{
library(ggplot2)
ggplot2:: ggplot(iris, aes(Sepal.Length, Sepal.Width))+
 geom_point()+
 theme_owid()
}
