% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_professional.R
\name{theme_professional_point}
\alias{theme_professional_point}
\title{theme_professional_point}
\usage{
theme_professional_point(
  base_size = 12,
  base_family = "Helvetica",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  border = FALSE
)
}
\arguments{
\item{border}{}
}
\value{

}
\description{
The theme_professional function sets a professional looking theme for your ggplot2 scatterplots.
Its minimalistic style doesn't distract from unnecessary grids and is easy to read.
For other plots, please use the theme_professional function.
}
