% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{theme_simple}
\usage{
theme_simple(yline = F)
}
\arguments{
\item{yline}{Add y line by setting to TRUE}
}
\value{

}
\description{
theme_simple
}
\examples{
library(ggplot2)
ggplot2:: ggplot(iris, aes(Sepal.Length, Sepal.Width))+
 geom_point()+
 theme_simple()

}
